refactor(work): split repository tests
